Moved config.mjs to src/web/ directory
