+++ src/web/config.mjs
